Implement user-specific features with persistence
Enabled features throughout the app dependent on the presence of a user by integrating a user context. This includes conditional rendering of the add link section only if a user is present, as well as storing user information in local storage for persistence across sessions. Introduced 'AddUserForm' to collect user data and dynamically display user details within the 'Avatar' component. These changes enhance user personalization and provide a foundation for future user-based functionality.

Refs #324, #112